requirements.txt
streamlit==1.45.1
langchain-mcp-adapters==0.1.7
langchain-openai==0.2.14
langchain-anthropic==0.3.15
langchain-google-genai==2.1.5
langchain-core==0.3.65
langgraph>=0.0.26
langgraph-checkpoint-sqlite>=1.0.0
aiosqlite>=0.17.0
mcp>=0.0.3
nest-asyncio>=1.6.0,<2.0.0
langchain-ollama==0.3.3
ollama>=0.4.8
jsonpointer==3.0.0
aiohttp==3.12.13
python-dotenv>=0.21.0
